[wip] build and deploy
